\docType{data}
\name{air}
\alias{air}
\title{Air data}
\format{A data frame with 111 rows and 4 columns
\cr \cr
Columns are:
\describe{
 \item{ozone}{surface concentration of ozone in New York,
in parts per million}
 \item{radiation}{solar radiation}
 \item{temperature}{observed temperature, in degrees Fahrenheit}
 \item{wind}{wind speed, in miles per hour}
}}
\source{
Chambers J, Hastie T (eds). 1992. \emph{Statistical Models
in S}. pg 348. Wadsworth and Brooks, Pacific Grove, CA.
}
\description{
Air data
}
\details{
A data frame with 111 observations (rows), and 4 variables
(columns), taken from an environmental study that measured
the four variables ozone, solar radiation, temperature, and
wind speed for 111 consecutive days.
}

