% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{d1}
\alias{d1}
\title{Dataset.}
\description{
Dataset.
}
\details{
This is documented to satisfy \code{R CMD check}.
\cr
I think this is required due to following lines from \code{\link{ci}}:
\cr
\code{d1 <- as.character(x$call$data)}
\cr
\code{c1 <- coxph(formula=f2, data=d1)}
}
\keyword{internal}

