\docType{package}
\name{survMisc-package}
\alias{survMisc}
\alias{survMisc-package}
\title{Miscellaneous functions for survival analysis.}
\description{
  Miscellaneous functions for survival analysis.
  \tabular{ll}{ Package: \tab survMisc\cr Type: \tab
  Package\cr Version: \tab 0.41\cr Date: \tab 2014-04-02\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes } A
  collection of functions for analysis of survival data.
  These extend the methods available in package survival.
}
\author{
  Chris Dardis \email{christopherdardis@gmail.com}
}
\keyword{package}

