\docType{data}
\name{tneKidney}
\alias{tneKidney}
\title{Time, no. at risk, no. events for kidney}
\format{A data frame with 16 rows and 7 columns}
\source{
  Klein J, Moeschberger. Survival Analysis, 2nd edition.
  Springer 2003. Example 7.9, pg 224.
}
\description{
  Data from

  Columns are: \describe{ \item{t}{time} \item{n_1}{no. at
  risk in group 1} \item{e_1}{no. events in group 1}
  \item{n_2}{no. at risk in group 2} \item{e_2}{no. events
  in group 2} \item{n}{no. at risk overall} \item{e}{no.
  events overall} }
}
\keyword{datasets}

