% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{air}
\alias{air}
\title{Environmental data timeseries}
\format{A \code{data.frame} with \eqn{111} observations (rows)
and \eqn{4} variables (columns).}
\source{
Chambers J, Hastie T (eds). 1992.
\emph{Statistical Models in S}. pg 348.
Wadsworth and Brooks, Pacific Grove, CA.
\href{http://dx.doi.org/10.1007/978-3-642-50096-1_48}{Springer (paywall)}
}
\description{
Environmental data timeseries
}
\details{
Data was taken from an environmental study
that measured the four variables on 111 consecutive days.
\cr \cr
Columns are:
\describe{
 \item{ozone}{surface concentration of ozone in New York,
              in parts per million}
 \item{radiation}{solar radiation}
 \item{temperature}{observed temperature, in degrees Fahrenheit}
 \item{wind}{wind speed, in miles per hour}
}
}
\seealso{
\code{\link{gamTerms}}
}
\keyword{datasets}

