% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{survMisc-package}
\alias{survMisc}
\alias{survMisc-package}
\title{Miscellaneous Functions for Survival Analysis.}
\description{
Miscellaneous Functions for Survival Analysis.
}
\details{
Miscellaneous functions for survival analysis.
\tabular{ll}{
 Package: \tab survMisc\cr
 Type: \tab Package\cr
  Version: \tab 0.4.4\cr
 Date: \tab 2014-12-21\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes
}
A collection of functions for the analysis of survival data. These
extend the methods already available in \code{package:survival}.
\cr
The intent is to generate a workspace for some of the common tasks
arising in survival analysis.
\cr
The package should be regarded as 'in development' until
release 1.0, meaning that there may be changes to certain function
names and parameters, although I will try to keep this to a minimum.
\cr
The package follows the \strong{camelCase} naming convention.
\cr \cr
For bug reports, feature requests or suggestions for improvement,
please try to submit to \href{https://github.com/dardisco/survMisc/issues}{github}.
Otherwise email me at the address below.
}
\author{
Chris Dardis \email{christopherdardis@gmail.com}
}
\concept{
survival
}
\keyword{package}

