% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{tneKidney}
\alias{tneKidney}
\title{Time, no. at risk, no. events for kidney data.}
\format{A data frame with 16 rows and 7 columns}
\source{
\bold{K&M}. Example 7.9, pg 224.
}
\description{
Data from Klein J and Moeschberger.
\cr
Columns are:
\describe{
 \item{t}{time}
 \item{n_1}{no. at risk in group 1}
 \item{e_1}{no. events in group 1}
 \item{n_2}{no. at risk in group 2}
 \item{e_2}{no. events in group 2}
 \item{n}{no. at risk overall}
 \item{e}{no. events overall}
}
}

